Patch for emscripten/wasm build with e9k debug graft
Four targeted fixes needed to build libretro-uae for the wasm
debugger backend:
- inputrecord.c/misc.c/sndboard.c: stub function signature fixes for
  emscripten linking
- memory.c: chip-RAM accessors call e9k_debug_memhook_* so
  watchpoints/protects work for the default A500/OCS no-cycle-exact
  config

+++ sources/src/sndboard.c
@@ -3161,7 +3161,7 @@ Exit:;
 #endif /* HAVE_X86 */
 #else
 
-void update_sndboard_sound (double clk) {}
+void update_sndboard_sound (float clk) {}
 void sndboard_ext_volume(void) {}
 
 struct autoconfig_info;
